pyproject.toml
[tool.poetry]
name = "cogstack-model-gateway"
version = "0.1.0"
description = "A Gateway for accessing Cogstack ModelServe instances"
authors = ["Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>"]
license = "Elastic License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
tenacity = "^9.0.0"
pika = "^1.3.2"
pydantic-settings = "^2.6.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
fastapi = {extras = ["standard"], version = "^0.115.5"}
docker = "^7.1.0"
mlflow = "^2.18.0"
python-multipart = "^0.0.19"
minio = "^7.2.12"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
ipykernel = "^6.29.5"
pre-commit = "^4.0.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
testcontainers = {extras = ["minio", "postgres", "rabbitmq"], version = "^4.9.0"}

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint.isort]
known-local-folder = ["common", "gateway", "scheduler"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
